Adds buildpack API to group.toml and metadata.toml
Now that we have buidpack API dependent launch logic, the launcher needs the api of each buildpack to exist in metadata.toml so it can do the right thing
* See #116 for an example (arg parsing for direct=false processes)

Other phases may have buidpack API dependent logic as well. Adding the `api` to group.toml will provide this the api of each buildpack everywhere it is needed.

Signed-off-by: Emily Casey <ecasey@vmware.com>